requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=9.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.25.0
gradio>=3.40.0
tensorflow>=2.13.0
mediapipe>=0.10.0
face-recognition>=1.3.0
dlib>=19.24.0
albumentations>=1.3.0
timm>=0.9.0
transformers>=4.30.0
wandb>=0.15.0
tqdm>=4.65.0
pandas>=2.0.0
plotly>=5.15.0
